Fixes #172 - Namespace for retry policies * Move RetryPolicies into its own Namespace * Fix Typo in ExponentialBackOffRetryPolicy
