+++ Funcky/PublicAPI.Shipped.txt
@@ -1,18 +1,6 @@
 #nullable enable
 
 ~Funcky.GenericConstraints.RequireClass<T>
-Funcky.ConstantDelayPolicy
-Funcky.ConstantDelayPolicy.ConstantDelayPolicy(int maxRetry, System.TimeSpan delay) -> void
-Funcky.ConstantDelayPolicy.Duration(int onRetryCount) -> System.TimeSpan
-Funcky.ConstantDelayPolicy.MaxRetries.get -> int
-Funcky.DoNotRetryPolicy
-Funcky.DoNotRetryPolicy.DoNotRetryPolicy() -> void
-Funcky.DoNotRetryPolicy.Duration(int onRetryCount) -> System.TimeSpan
-Funcky.DoNotRetryPolicy.MaxRetries.get -> int
-Funcky.ExponentialBackoffRetryPolicy
-Funcky.ExponentialBackoffRetryPolicy.Duration(int onRetryCount) -> System.TimeSpan
-Funcky.ExponentialBackoffRetryPolicy.ExponentialBackoffRetryPolicy(int maxRetry, System.TimeSpan firstDelay) -> void
-Funcky.ExponentialBackoffRetryPolicy.MaxRetries.get -> int
 Funcky.Extensions.AsyncEnumerableExtensions
 Funcky.Extensions.DictionaryExtensions
 Funcky.Extensions.EnumerableExtensions
@@ -41,13 +29,6 @@ Funcky.Functional
 Funcky.GenericConstraints.RequireClass<T>.RequireClass() -> void
 Funcky.GenericConstraints.RequireStruct<T>
 Funcky.GenericConstraints.RequireStruct<T>.RequireStruct() -> void
-Funcky.IRetryPolicy
-Funcky.IRetryPolicy.Duration(int onRetryCount) -> System.TimeSpan
-Funcky.IRetryPolicy.MaxRetries.get -> int
-Funcky.LinearBackoffRetryPolicy
-Funcky.LinearBackoffRetryPolicy.Duration(int onRetryCount) -> System.TimeSpan
-Funcky.LinearBackoffRetryPolicy.LinearBackoffRetryPolicy(int maxRetry, System.TimeSpan firstDelay) -> void
-Funcky.LinearBackoffRetryPolicy.MaxRetries.get -> int
 Funcky.Monads.AwaitableOptionExtensions
 Funcky.Monads.Either<TLeft, TRight>
 Funcky.Monads.Either<TLeft, TRight>.Equals(Funcky.Monads.Either<TLeft, TRight> other) -> bool
@@ -93,8 +74,6 @@ Funcky.Monads.Result<TValidResult>.Match(System.Action<TValidResult>! ok, System
 Funcky.Monads.Result<TValidResult>.Match<TMatchResult>(System.Func<TValidResult, TMatchResult>! ok, System.Func<System.Exception!, TMatchResult>! error) -> TMatchResult
 Funcky.Monads.Result<TValidResult>.Select<TResult>(System.Func<TValidResult, TResult>! selector) -> Funcky.Monads.Result<TResult>
 Funcky.Monads.Result<TValidResult>.SelectMany<TSelectedResult, TResult>(System.Func<TValidResult, Funcky.Monads.Result<TSelectedResult>>! selectedResultSelector, System.Func<TValidResult, TSelectedResult, TResult>! resultSelector) -> Funcky.Monads.Result<TResult>
-Funcky.NoDelayRetryPolicy
-Funcky.NoDelayRetryPolicy.NoDelayRetryPolicy(int maxRetry) -> void
 Funcky.Sequence
 Funcky.Unit
 Funcky.Unit.CompareTo(Funcky.Unit other) -> int
@@ -297,8 +276,6 @@ static Funcky.Functional.NoOperation<T1, T2>(T1 p1, T2 p2) -> void
 static Funcky.Functional.NoOperation<T1>(T1 p1) -> void
 static Funcky.Functional.Not<T>(System.Func<T, bool>! predicate) -> System.Func<T, bool>!
 static Funcky.Functional.Retry<TResult>(System.Func<Funcky.Monads.Option<TResult>>! producer) -> TResult
-static Funcky.Functional.Retry<TResult>(System.Func<Funcky.Monads.Option<TResult>>! producer, Funcky.IRetryPolicy! retryPolicy) -> Funcky.Monads.Option<TResult>
-static Funcky.Functional.RetryAsync<TResult>(System.Func<System.Threading.Tasks.Task<Funcky.Monads.Option<TResult>>!>! producer, Funcky.IRetryPolicy! retryPolicy) -> System.Threading.Tasks.Task<Funcky.Monads.Option<TResult>>!
 static Funcky.Functional.True<T>(T ω) -> bool
 static Funcky.Functional.Uncurry<T1, T2, T3, T4, T5, T6, T7, T8, TResult>(System.Func<T1, System.Func<T2, System.Func<T3, System.Func<T4, System.Func<T5, System.Func<T6, System.Func<T7, System.Func<T8, TResult>!>!>!>!>!>!>!>! function) -> System.Func<T1, T2, T3, T4, T5, T6, T7, T8, TResult>!
 static Funcky.Functional.Uncurry<T1, T2, T3, T4, T5, T6, T7, TResult>(System.Func<T1, System.Func<T2, System.Func<T3, System.Func<T4, System.Func<T5, System.Func<T6, System.Func<T7, TResult>!>!>!>!>!>!>! function) -> System.Func<T1, T2, T3, T4, T5, T6, T7, TResult>!
