+++ Funcky/RetryPolicies/DoNotRetryPolicy.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Funcky
+namespace Funcky.RetryPolicies
 {
     public sealed class DoNotRetryPolicy : IRetryPolicy
     {
